Modify GetIterator to return a generator
- Fixes #2
- Also fixes a failing test

Calling GetIterator originally converted itself to an array recursively
which causes any nested Maps to be converted to arrays as well. This is
not how an iterator should work. By using a generator we can return the
actual values without mutating them.

A side effect of this is that the new GetIterator method no longer
returns a "copy" of the Map. I can't see any downside of this since the
caller shouldn't be modifying an iterable inside a loop anyway.